// frontend/plugins/welcome/src/api/models/EntPromotion.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SUT SA Example API Patient
 * This is a sample server for SUT SE 2563
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    EntPromotionEdges,
    EntPromotionEdgesFromJSON,
    EntPromotionEdgesFromJSONTyped,
    EntPromotionEdgesToJSON,
} from './';

/**
 * 
 * @export
 * @interface EntPromotion
 */
export interface EntPromotion {
    /**
     * CODE holds the value of the "CODE" field.
     * @type {string}
     * @memberof EntPromotion
     */
    cODE?: string;
    /**
     * DATE holds the value of the "DATE" field.
     * @type {string}
     * @memberof EntPromotion
     */
    dATE?: string;
    /**
     * DESC holds the value of the "DESC" field.
     * @type {string}
     * @memberof EntPromotion
     */
    dESC?: string;
    /**
     * NAME holds the value of the "NAME" field.
     * @type {string}
     * @memberof EntPromotion
     */
    nAME?: string;
    /**
     * 
     * @type {EntPromotionEdges}
     * @memberof EntPromotion
     */
    edges?: EntPromotionEdges;
    /**
     * ID of the ent.
     * @type {number}
     * @memberof EntPromotion
     */
    id?: number;
}

export function EntPromotionFromJSON(json: any): EntPromotion {
    return EntPromotionFromJSONTyped(json, false);
}

export function EntPromotionFromJSONTyped(json: any, ignoreDiscriminator: boolean): EntPromotion {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'cODE': !exists(json, 'CODE') ? undefined : json['CODE'],
        'dATE': !exists(json, 'DATE') ? undefined : json['DATE'],
        'dESC': !exists(json, 'DESC') ? undefined : json['DESC'],
        'nAME': !exists(json, 'NAME') ? undefined : json['NAME'],
        'edges': !exists(json, 'edges') ? undefined : EntPromotionEdgesFromJSON(json['edges']),
        'id': !exists(json, 'id') ? undefined : json['id'],
    };
}

export function EntPromotionToJSON(value?: EntPromotion | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'CODE': value.cODE,
        'DATE': value.dATE,
        'DESC': value.dESC,
        'NAME': value.nAME,
        'edges': EntPromotionEdgesToJSON(value.edges),
        'id': value.id,
    };
}